+++ commands/process.js
@@ -19,23 +19,32 @@ function process (options) {
 
     console.log(chalk.gray.bold('File name          : ' + options.name ))
     console.log(chalk.gray.bold('File path          : ' + options.destination))
-    console.log(chalk.gray.bold('File origin path   : ' + options.file))
+    console.log(chalk.gray.bold('Files to process   : '))
+    options.file.forEach(file => {
+        console.log(chalk.gray.bold('- ' + file))
+    })
     console.log(chalk.gray.bold('CSS theme          : ' + options.theme))
 
-    fs.readFile(options.file, {encoding: 'utf-8'}, function (err, data) {
-        if (!err) {
-            htmlProcessor.generateHtmlFile(
-                options.titre,
-                options.destination + options.name,
-                data,
-                options.theme
-            )
-
-            console.log(chalk.blue.bold('Generated file with success :: ' + options.destination + options.name))
-        } else {
-            // console.log(err);
-            console.log(chalk.red.bold("File " + options.file + " doesn't exist"))
+    let datas = ""
+    let continueProcessing = true
+    options.file.forEach(file_path => {
+        try {
+            datas = datas + fs.readFileSync(file_path, {encoding: 'utf-8'}) + "\n"
+        } catch (e) {
+            continueProcessing = false
+            console.log(chalk.red.bold("File " + file_path + " doesn't exist"))
         }
-    });
+    })
+
+    if (continueProcessing) {
+        htmlProcessor.generateHtmlFile(
+            options.titre,
+            options.destination + options.name,
+            datas,
+            options.theme
+        )
+
+        console.log(chalk.blue.bold('Generated file with success :: ' + options.destination + options.name))
+    }
 }
 module.exports = process